Updated Vision settings to allow screenshots for desktop through SKSE_HTTP
This modification requires the modified SKSE_HTTP with screenshot taking functions.

The Vision Settings have been updated to automatically offer the in-engine screenshot taking for any Skyrim desktop version.
Vision script is updated for in-engine screenshot taking for desktop.
Minor changes to MantellaConstants and MantellaConversation to interact properly with later versions of Mantella Software.